Unification of LinearEntity and LinearEntity3D
