+++ sympy/geometry/entity.py
@@ -35,12 +35,15 @@ ordering_of_classes = [
     "Point2D",
     "Point3D",
     "Point",
+    "Segment2D",
+    "Ray2D",
+    "Line2D",
+    "Segment3D",
+    "Line3D",
+    "Ray3D",
     "Segment",
     "Ray",
     "Line",
-    "Line3D",
-    "Ray3D",
-    "Segment3D",
     "Plane",
     "Triangle",
     "RegularPolygon",
