+++ sympy/geometry/line.py
@@ -6,6 +6,14 @@ LinearEntity
 Line
 Ray
 Segment
+LinearEntity2D
+Line2D
+Ray2D
+Segment2D
+LinearEntity3D
+Line3D
+Ray3D
+Segment3D
 
 """
 from __future__ import division, print_function
@@ -25,9 +33,10 @@ from sympy.sets import Intersection, EmptySet
 from sympy.matrices import Matrix
 
 from .entity import GeometryEntity, GeometrySet
-from .point import Point
+from .point import Point, Point3D
 from .util import _symbol
 
+import warnings
 
 # TODO: this should be placed elsewhere and reused in other modules
 
@@ -37,16 +46,14 @@ class Undecidable(ValueError):
 
 
 class LinearEntity(GeometrySet):
-    """A base class for all linear entities (line, ray and segment)
-    in a 2-dimensional Euclidean space.
+    """A base class for all linear entities (Line, Ray and Segment)
+    in n-dimensional Euclidean space.
 
     Attributes
     ==========
 
     p1
     p2
-    coefficients
-    slope
     points
 
     Notes
@@ -203,9 +210,17 @@ class LinearEntity(GeometrySet):
         >>> l1, l2, l3 = Line(p1, p2), Line(p1, p3), Line(p1, p4)
         >>> Line.are_concurrent(l1, l2, l3)
         True
-
         >>> l4 = Line(p2, p3)
         >>> Line.are_concurrent(l2, l3, l4)
+        False
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2 = Point3D(0, 0, 0), Point3D(3, 5, 2)
+        >>> p3, p4 = Point3D(-2, -2, -2), Point3D(0, 2, 1)
+        >>> l1, l2, l3 = Line3D(p1, p2), Line3D(p1, p3), Line3D(p1, p4)
+        >>> Line3D.are_concurrent(l1, l2, l3)
+        True
+        >>> l4 = Line3D(p2, p3)
+        >>> Line3D.are_concurrent(l2, l3, l4)
         False
 
         """
@@ -258,10 +273,19 @@ class LinearEntity(GeometrySet):
         >>> l1, l2 = Line(p1, p2), Line(p3, p4)
         >>> Line.is_parallel(l1, l2)
         True
-
         >>> p5 = Point(6, 6)
         >>> l3 = Line(p3, p5)
         >>> Line.is_parallel(l1, l3)
+        False
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2 = Point3D(0, 0, 0), Point3D(3, 4, 5)
+        >>> p3, p4 = Point3D(2, 1, 1), Point3D(8, 9, 11)
+        >>> l1, l2 = Line3D(p1, p2), Line3D(p3, p4)
+        >>> Line3D.is_parallel(l1, l2)
+        True
+        >>> p5 = Point3D(6, 6, 6)
+        >>> l3 = Line3D(p3, p5)
+        >>> Line3D.is_parallel(l1, l3)
         False
 
         """
@@ -299,9 +323,17 @@ class LinearEntity(GeometrySet):
         >>> l1, l2 = Line(p1, p2), Line(p1, p3)
         >>> l1.is_perpendicular(l2)
         True
-
         >>> p4 = Point(5, 3)
         >>> l3 = Line(p1, p4)
+        >>> l1.is_perpendicular(l3)
+        False
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(1, 1, 1), Point3D(-1, 2, 0)
+        >>> l1, l2 = Line3D(p1, p2), Line3D(p2, p3)
+        >>> l1.is_perpendicular(l2)
+        False
+        >>> p4 = Point3D(5, 3, 7)
+        >>> l3 = Line3D(p1, p4)
         >>> l1.is_perpendicular(l3)
         False
 
@@ -349,6 +381,11 @@ class LinearEntity(GeometrySet):
         >>> l1, l2 = Line(p1, p2), Line(p1, p3)
         >>> l1.angle_between(l2)
         pi/2
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(1, 1, 1), Point3D(-1, 2, 0)
+        >>> l1, l2 = Line3D(p1, p2), Line3D(p2, p3)
+        >>> l1.angle_between(l2)
+        acos(-sqrt(2)/3)
 
         """
         if not isinstance(l1, LinearEntity) and not isinstance(l2, LinearEntity):
@@ -382,6 +419,14 @@ class LinearEntity(GeometrySet):
         >>> from sympy import Point, Line
         >>> p1, p2, p3 = Point(0, 0), Point(2, 3), Point(-2, 2)
         >>> l1 = Line(p1, p2)
+        >>> l2 = l1.parallel_line(p3)
+        >>> p3 in l2
+        True
+        >>> l1.is_parallel(l2)
+        True
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(2, 3, 4), Point3D(-2, 2, 0)
+        >>> l1 = Line3D(p1, p2)
         >>> l2 = l1.parallel_line(p3)
         >>> p3 in l2
         True
@@ -422,10 +467,20 @@ class LinearEntity(GeometrySet):
         True
         >>> l1.is_perpendicular(l2)
         True
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(2, 3, 4), Point3D(-2, 2, 0)
+        >>> l1 = Line3D(p1, p2)
+        >>> l2 = l1.perpendicular_line(p3)
+        >>> p3 in l2
+        True
+        >>> l1.is_perpendicular(l2)
+        True
 
         """
         p = Point(p)
-        return Line(p, p + self.direction.orthogonal_direction)
+        if p in self:
+            p = p + self.direction.orthogonal_direction
+        return Line(p, self.projection(p))
 
     def perpendicular_segment(self, p):
         """Create a perpendicular line segment from `p` to this line.
@@ -466,7 +521,17 @@ class LinearEntity(GeometrySet):
         >>> p3 in s1
         True
         >>> l1.perpendicular_segment(Point(4, 0))
-        Segment(Point2D(2, 2), Point2D(4, 0))
+        Segment2D(Point2D(2, 2), Point2D(4, 0))
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(1, 1, 1), Point3D(0, 2, 0)
+        >>> l1 = Line3D(p1, p2)
+        >>> s1 = l1.perpendicular_segment(p3)
+        >>> l1.is_perpendicular(s1)
+        True
+        >>> p3 in s1
+        True
+        >>> l1.perpendicular_segment(Point3D(4, 0, 0))
+        Segment3D(Point3D(4/3, 4/3, 4/3), Point3D(4, 0, 0))
 
         """
         p = Point(p)
@@ -494,40 +559,6 @@ class LinearEntity(GeometrySet):
         return S.Infinity
 
     @property
-    def slope(self):
-        """The slope of this linear entity, or infinity if vertical.
-
-        Returns
-        =======
-
-        slope : number or sympy expression
-
-        See Also
-        ========
-
-        coefficients
-
-        Examples
-        ========
-
-        >>> from sympy import Point, Line
-        >>> p1, p2 = Point(0, 0), Point(3, 5)
-        >>> l1 = Line(p1, p2)
-        >>> l1.slope
-        5/3
-
-        >>> p3 = Point(0, 4)
-        >>> l2 = Line(p1, p3)
-        >>> l2.slope
-        oo
-
-        """
-        d1, d2 = (self.p1 - self.p2).args
-        if d1 == 0:
-            return S.Infinity
-        return simplify(d2/d1)
-
-    @property
     def points(self):
         """The two points used to define this linear entity.
 
@@ -552,18 +583,6 @@ class LinearEntity(GeometrySet):
 
         """
         return (self.p1, self.p2)
-
-    @property
-    def bounds(self):
-        """Return a tuple (xmin, ymin, xmax, ymax) representing the bounding
-        rectangle for the geometric figure.
-
-        """
-
-        verts = self.points
-        xs = [p.x for p in verts]
-        ys = [p.y for p in verts]
-        return (min(xs), min(ys), max(xs), max(ys))
 
     def projection(self, other):
         """Project a point, line, ray, or segment onto this linear entity.
@@ -609,13 +628,23 @@ class LinearEntity(GeometrySet):
         >>> l1 = Line(p1, p2)
         >>> l1.projection(p3)
         Point2D(1/4, 1/4)
-
         >>> p4, p5 = Point(10, 0), Point(12, 1)
         >>> s1 = Segment(p4, p5)
         >>> l1.projection(s1)
-        Segment(Point2D(5, 5), Point2D(13/2, 13/2))
+        Segment2D(Point2D(5, 5), Point2D(13/2, 13/2))
+        >>> p1, p2, p3 = Point(0, 0, 1), Point(1, 1, 2), Point(2, 0, 1)
+        >>> l1 = Line(p1, p2)
+        >>> l1.projection(p3)
+        Point3D(2/3, 2/3, 5/3)
+        >>> p4, p5 = Point(10, 0, 1), Point(12, 1, 3)
+        >>> s1 = Segment(p4, p5)
+        >>> l1.projection(s1)
+        Segment3D(Point3D(10/3, 10/3, 13/3), Point3D(5, 5, 6))
 
         """
+
+        if is_sequence(other):
+            other = Point(other)
 
         def proj_point(p):
             return Point.project(p - self.p1, self.direction) + self.p1
@@ -664,14 +693,25 @@ class LinearEntity(GeometrySet):
         >>> l1 = Line(p1, p2)
         >>> l1.intersection(p3)
         [Point2D(7, 7)]
-
         >>> p4, p5 = Point(5, 0), Point(0, 3)
         >>> l2 = Line(p4, p5)
         >>> l1.intersection(l2)
         [Point2D(15/8, 15/8)]
-
         >>> p6, p7 = Point(0, 5), Point(2, 6)
         >>> s1 = Segment(p6, p7)
+        >>> l1.intersection(s1)
+        []
+        >>> from sympy import Point3D, Line3D, Segment3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(1, 1, 1), Point3D(7, 7, 7)
+        >>> l1 = Line3D(p1, p2)
+        >>> l1.intersection(p3)
+        [Point3D(7, 7, 7)]
+        >>> l1 = Line3D(Point3D(4,19,12), Point3D(5,25,17))
+        >>> l2 = Line3D(Point3D(-3, -15, -19), direction_ratio=[2,8,8])
+        >>> l1.intersection(l2)
+        [Point3D(1, 1, -3)]
+        >>> p6, p7 = Point3D(0, 5, 2), Point3D(2, 6, 3)
+        >>> s1 = Segment3D(p6, p7)
         >>> l1.intersection(s1)
         []
 
@@ -805,6 +845,11 @@ class LinearEntity(GeometrySet):
         >>> l1 = Line(p1, p2)
         >>> l1.arbitrary_point()
         Point2D(4*t + 1, 3*t)
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2 = Point3D(1, 0, 0), Point3D(5, 3, 1)
+        >>> l1 = Line3D(p1, p2)
+        >>> l1.arbitrary_point()
+        Point3D(4*t + 1, 3*t, t)
 
         """
         t = _symbol(parameter)
@@ -908,30 +953,35 @@ class LinearEntity(GeometrySet):
             return 1
         return -1
 
-
 class Line(LinearEntity):
     """An infinite line in space.
 
-    A line is declared with two distinct points or a point and slope
-    as defined using keyword `slope`.
-
-    Notes
-    =====
-
-    At the moment only lines in a 2D space can be declared, because
-    Points can be defined only for 2D spaces.
+    A line is declared with two distinct points.
+    A 2D line may be declared with a point and slope
+    and a 3D line may be defined with a point and a direction ratio.
 
     Parameters
     ==========
 
     p1 : Point
-    pt : Point
+    p2 : Point
     slope : sympy expression
+    direction_ratio : list
+
+    Notes
+    =====
+
+    `Line` will automatically subclass to `Line2D` or `Line3D` based
+    on the dimension of `p1`.  The `slope` argument is only relevant
+    for `Line2D` and the `direction_ratio` argument is only relevant
+    for `Line3D`.
 
     See Also
     ========
 
     sympy.geometry.point.Point
+    sympy.geometry.line.Line2D
+    sympy.geometry.line.Line3D
 
     Examples
     ========
@@ -942,7 +992,7 @@ class Line(LinearEntity):
     >>> from sympy.geometry import Line, Segment
     >>> L = Line(Point(2,3), Point(3,5))
     >>> L
-    Line(Point2D(2, 3), Point2D(3, 5))
+    Line2D(Point2D(2, 3), Point2D(3, 5))
     >>> L.points
     (Point2D(2, 3), Point2D(3, 5))
     >>> L.equation()
@@ -953,7 +1003,7 @@ class Line(LinearEntity):
     Instantiate with keyword ``slope``:
 
     >>> Line(Point(0, 0), slope=0)
-    Line(Point2D(0, 0), Point2D(1, 0))
+    Line2D(Point2D(0, 0), Point2D(1, 0))
 
     Instantiate with another linear object
 
@@ -962,33 +1012,17 @@ class Line(LinearEntity):
     x
     """
 
-    def __new__(cls, p1, pt=None, slope=None, **kwargs):
+    def __new__(cls, p1, *pts, **kwargs):
         if isinstance(p1, LinearEntity):
-            p1, pt = p1.args
+            ambient_dim = len(p1.p1)
         else:
-            p1 = Point(p1)
-        if pt is not None and slope is None:
-            try:
-                p2 = Point(pt)
-            except NotImplementedError:
-                raise ValueError('The 2nd argument was not a valid Point. '
-                'If it was a slope, enter it with keyword "slope".')
-        elif slope is not None and pt is None:
-            slope = sympify(slope)
-            if slope.is_finite is False:
-                # when infinite slope, don't change x
-                dx = 0
-                dy = 1
-            else:
-                # go over 1 up slope
-                dx = 1
-                dy = slope
-            # XXX avoiding simplification by adding to coords directly
-            p2 = Point(p1.x + dx, p1.y + dy)
-        else:
-            raise ValueError('A 2nd Point or keyword "slope" must be used.')
+            ambient_dim = len(p1)
 
-        return LinearEntity.__new__(cls, p1, p2, **kwargs)
+        if ambient_dim == 2:
+            return Line2D(p1, *pts, **kwargs)
+        elif ambient_dim == 3:
+            return Line3D(p1, *pts, **kwargs)
+        return LinearEntity.__new__(cls, p1, *pts, **kwargs)
 
     def plot_interval(self, parameter='t'):
         """The plot interval for the default geometric plot of line. Gives
@@ -1020,47 +1054,6 @@ class Line(LinearEntity):
         t = _symbol(parameter)
         return [t, -5, 5]
 
-    def equation(self, x='x', y='y'):
-        """The equation of the line: ax + by + c.
-
-        Parameters
-        ==========
-
-        x : str, optional
-            The name to use for the x-axis, default value is 'x'.
-        y : str, optional
-            The name to use for the y-axis, default value is 'y'.
-
-        Returns
-        =======
-
-        equation : sympy expression
-
-        See Also
-        ========
-
-        LinearEntity.coefficients
-
-        Examples
-        ========
-
-        >>> from sympy import Point, Line
-        >>> p1, p2 = Point(1, 0), Point(5, 3)
-        >>> l1 = Line(p1, p2)
-        >>> l1.equation()
-        -3*x + 4*y + 3
-
-        """
-        x, y = _symbol(x), _symbol(y)
-        p1, p2 = self.points
-        if p1.x == p2.x:
-            return x - p1.x
-        elif p1.y == p2.y:
-            return y - p1.y
-
-        a, b, c = self.coefficients
-        return a*x + b*y + c
-
     def contains(self, other):
         """
         Return True if `other` is on this Line, or False otherwise.
@@ -1077,9 +1070,24 @@ class Line(LinearEntity):
         True
         >>> l.contains((0, 0))
         False
+        >>> a = (0, 0, 0)
+        >>> b = (1, 1, 1)
+        >>> c = (2, 2, 2)
+        >>> l1 = Line(a, b)
+        >>> l2 = Line(b, a)
+        >>> l1 == l2
+        False
+        >>> l1 in l2
+        True
+
         """
         if is_sequence(other):
             other = Point(other)
+
+        if hasattr(other, 'ambient_dimension') and other.ambient_dimension != self.ambient_dimension:
+            warnings.warn("Testing containment for objects of different ambient dimensions", RuntimeWarning)
+            return False
+
         if isinstance(other, Point):
             return Point.is_collinear(other, self.p1, self.p2)
         if isinstance(other, LinearEntity):
@@ -1093,7 +1101,7 @@ class Line(LinearEntity):
         Raises
         ======
 
-        NotImplementedError is raised if o is not a Point
+        NotImplementedError is raised if `other` is not a Point
 
         Examples
         ========
@@ -1105,11 +1113,20 @@ class Line(LinearEntity):
         sqrt(2)
         >>> s.distance((-1, 2))
         3*sqrt(2)/2
+        >>> p1, p2 = Point(0, 0, 0), Point(1, 1, 1)
+        >>> s = Line(p1, p2)
+        >>> s.distance(Point(-1, 1, 1))
+        2*sqrt(6)/3
+        >>> s.distance((-1, 1, 1))
+        2*sqrt(6)/3
+
         """
         if is_sequence(other):
             other = Point(other)
         if not isinstance(other, Point):
             raise TypeError("`other` must be a Point, not {}".format(other))
+        if self.contains(other):
+            return S.Zero
         return self.perpendicular_segment(other).length
 
     @deprecated(useinstead="equals", deprecated_since_version="1.0")
@@ -1122,33 +1139,8 @@ class Line(LinearEntity):
             return False
         return Point.is_collinear(self.p1, other.p1, self.p2, other.p2)
 
-    def _svg(self, scale_factor=1., fill_color="#66cc99"):
-        """Returns SVG path element for the LinearEntity.
-
-        Parameters
-        ==========
-
-        scale_factor : float
-            Multiplication factor for the SVG stroke-width.  Default is 1.
-        fill_color : str, optional
-            Hex string for fill color. Default is "#66cc99".
-        """
-
-        from sympy.core.evalf import N
-
-        verts = (N(self.p1), N(self.p2))
-        coords = ["{0},{1}".format(p.x, p.y) for p in verts]
-        path = "M {0} L {1}".format(coords[0], " L ".join(coords[1:]))
-
-        return (
-            '<path fill-rule="evenodd" fill="{2}" stroke="#555555" '
-            'stroke-width="{0}" opacity="0.6" d="{1}" '
-            'marker-start="url(#markerReverseArrow)" marker-end="url(#markerArrow)"/>'
-            ).format(2. * scale_factor, path, fill_color)
-
 class Ray(LinearEntity):
-    """
-    A Ray is a semi-line in the space with a source point and a direction.
+    """A Ray is a semi-line in the space with a source point and a direction.
 
     Parameters
     ==========
@@ -1164,19 +1156,20 @@ class Ray(LinearEntity):
     ==========
 
     source
-    xdirection
-    ydirection
 
     See Also
     ========
 
-    sympy.geometry.point.Point, Line
+    sympy.geometry.line.Ray2D
+    sympy.geometry.line.Ray3D
+    sympy.geometry.point.Point
+    sympy.geometry.line.Line
 
     Notes
     =====
 
-    At the moment only rays in a 2D space can be declared, because
-    Points can be defined only for 2D spaces.
+    `Ray` will automatically subclass to `Ray2D` or `Ray3D` based on the
+    dimension of `p1`.
 
     Examples
     ========
@@ -1188,7 +1181,7 @@ class Ray(LinearEntity):
     >>> r = Ray(Point(2, 3), Point(3, 5))
     >>> r = Ray(Point(2, 3), Point(3, 5))
     >>> r
-    Ray(Point2D(2, 3), Point2D(3, 5))
+    Ray2D(Point2D(2, 3), Point2D(3, 5))
     >>> r.points
     (Point2D(2, 3), Point2D(3, 5))
     >>> r.source
@@ -1203,50 +1196,17 @@ class Ray(LinearEntity):
     1
 
     """
-
-    def __new__(cls, p1, pt=None, angle=None, **kwargs):
-        p1 = Point(p1)
-        if pt is not None and angle is None:
-            try:
-                p2 = Point(pt)
-            except NotImplementedError:
-                from sympy.utilities.misc import filldedent
-                raise ValueError(filldedent('''
-                    The 2nd argument was not a valid Point; if
-                    it was meant to be an angle it should be
-                    given with keyword "angle".'''))
-            if p1 == p2:
-                raise ValueError('A Ray requires two distinct points.')
-        elif angle is not None and pt is None:
-            # we need to know if the angle is an odd multiple of pi/2
-            c = pi_coeff(sympify(angle))
-            p2 = None
-            if c is not None:
-                if c.is_Rational:
-                    if c.q == 2:
-                        if c.p == 1:
-                            p2 = p1 + Point(0, 1)
-                        elif c.p == 3:
-                            p2 = p1 + Point(0, -1)
-                    elif c.q == 1:
-                        if c.p == 0:
-                            p2 = p1 + Point(1, 0)
-                        elif c.p == 1:
-                            p2 = p1 + Point(-1, 0)
-                if p2 is None:
-                    c *= S.Pi
-            else:
-                c = angle % (2*S.Pi)
-            if not p2:
-                m = 2*c/S.Pi
-                left = And(1 < m, m < 3)  # is it in quadrant 2 or 3?
-                x = Piecewise((-1, left), (Piecewise((0, Eq(m % 1, 0)), (1, True)), True))
-                y = Piecewise((-tan(c), left), (Piecewise((1, Eq(m, 1)), (-1, Eq(m, 3)), (tan(c), True)), True))
-                p2 = p1 + Point(x, y)
+    def __new__(cls, p1, *pts, **kwargs):
+        if isinstance(p1, LinearEntity):
+            ambient_dim = len(p1.p1)
         else:
-            raise ValueError('A 2nd point or keyword "angle" must be used.')
+            ambient_dim = len(p1)
 
-        return LinearEntity.__new__(cls, p1, p2, **kwargs)
+        if ambient_dim == 2:
+            return Ray2D(p1, *pts, **kwargs)
+        elif ambient_dim == 3:
+            return Ray3D(p1, *pts, **kwargs)
+        return LinearEntity.__new__(cls, p1, *pts, **kwargs)
 
     @property
     def source(self):
@@ -1265,74 +1225,13 @@ class Ray(LinearEntity):
         >>> r1 = Ray(p1, p2)
         >>> r1.source
         Point2D(0, 0)
+        >>> p1, p2 = Point(0, 0, 0), Point(4, 1, 5)
+        >>> r1 = Ray(p2, p1)
+        >>> r1.source
+        Point3D(4, 1, 5)
 
         """
         return self.p1
-
-    @property
-    def xdirection(self):
-        """The x direction of the ray.
-
-        Positive infinity if the ray points in the positive x direction,
-        negative infinity if the ray points in the negative x direction,
-        or 0 if the ray is vertical.
-
-        See Also
-        ========
-
-        ydirection
-
-        Examples
-        ========
-
-        >>> from sympy import Point, Ray
-        >>> p1, p2, p3 = Point(0, 0), Point(1, 1), Point(0, -1)
-        >>> r1, r2 = Ray(p1, p2), Ray(p1, p3)
-        >>> r1.xdirection
-        oo
-        >>> r2.xdirection
-        0
-
-        """
-        if self.p1.x < self.p2.x:
-            return S.Infinity
-        elif self.p1.x == self.p2.x:
-            return S.Zero
-        else:
-            return S.NegativeInfinity
-
-    @property
-    def ydirection(self):
-        """The y direction of the ray.
-
-        Positive infinity if the ray points in the positive y direction,
-        negative infinity if the ray points in the negative y direction,
-        or 0 if the ray is horizontal.
-
-        See Also
-        ========
-
-        xdirection
-
-        Examples
-        ========
-
-        >>> from sympy import Point, Ray
-        >>> p1, p2, p3 = Point(0, 0), Point(-1, -1), Point(-1, 0)
-        >>> r1, r2 = Ray(p1, p2), Ray(p1, p3)
-        >>> r1.ydirection
-        -oo
-        >>> r2.ydirection
-        0
-
-        """
-        if self.p1.y < self.p2.y:
-            return S.Infinity
-        elif self.p1.y == self.p2.y:
-            return S.Zero
-        else:
-            return S.NegativeInfinity
-
 
     def distance(self, other):
         """
@@ -1341,7 +1240,7 @@ class Ray(LinearEntity):
         Raises
         ======
 
-        NotImplementedError is raised if o is not a Point
+        NotImplementedError is raised if `other` is not a Point
 
         Examples
         ========
@@ -1353,6 +1252,15 @@ class Ray(LinearEntity):
         sqrt(2)
         >>> s.distance((-1, 2))
         3*sqrt(2)/2
+        >>> p1, p2 = Point(0, 0, 0), Point(1, 1, 2)
+        >>> s = Ray(p1, p2)
+        >>> s
+        Ray3D(Point3D(0, 0, 0), Point3D(1, 1, 2))
+        >>> s.distance(Point(-1, -1, 2))
+        4*sqrt(3)/3
+        >>> s.distance((-1, -1, 2))
+        4*sqrt(3)/3
+
         """
         other = Point(other)
         if self.contains(other):
@@ -1431,6 +1339,10 @@ class Ray(LinearEntity):
         if is_sequence(other):
             other = Point(other)
 
+        if hasattr(other, 'ambient_dimension') and other.ambient_dimension != self.ambient_dimension:
+            warnings.warn("Testing containment for objects of different ambient dimensions", RuntimeWarning)
+            return False
+
         if isinstance(other, Point):
             if Point.is_collinear(self.p1, self.p2, other):
                 # if we're in the direction of the ray, our
@@ -1472,7 +1384,6 @@ class Ray(LinearEntity):
             'marker-start="url(#markerCircle)" marker-end="url(#markerArrow)"/>'
             ).format(2. * scale_factor, path, fill_color)
 
-
 class Segment(LinearEntity):
     """An undirected line segment in space.
 
@@ -1491,13 +1402,16 @@ class Segment(LinearEntity):
     See Also
     ========
 
-    sympy.geometry.point.Point, Line
+    sympy.geometry.line.Segment2D
+    sympy.geometry.line.Segment3D
+    sympy.geometry.point.Point
+    sympy.geometry.line.Line
 
     Notes
     =====
 
-    At the moment only segments in a 2D space can be declared, because
-    Points can be defined only for 2D spaces.
+    If 2D or 3D points are used to define `Segment`, it will
+    be automatically subclassed to `Segment2D` or `Segment3D`.
 
     Examples
     ========
@@ -1507,10 +1421,10 @@ class Segment(LinearEntity):
     >>> from sympy.abc import s
     >>> from sympy.geometry import Segment
     >>> Segment((1, 0), (1, 1)) # tuples are interpreted as pts
-    Segment(Point2D(1, 0), Point2D(1, 1))
+    Segment2D(Point2D(1, 0), Point2D(1, 1))
     >>> s = Segment(Point(4, 3), Point(1, 1))
     >>> s
-    Segment(Point2D(1, 1), Point2D(4, 3))
+    Segment2D(Point2D(1, 1), Point2D(4, 3))
     >>> s.points
     (Point2D(1, 1), Point2D(4, 3))
     >>> s.slope
@@ -1519,22 +1433,30 @@ class Segment(LinearEntity):
     sqrt(13)
     >>> s.midpoint
     Point2D(5/2, 2)
+    >>> Segment((1, 0, 0), (1, 1, 1)) # tuples are interpreted as pts
+    Segment3D(Point3D(1, 0, 0), Point3D(1, 1, 1))
+    >>> s = Segment(Point(4, 3, 9), Point(1, 1, 7))
+    >>> s
+    Segment3D(Point3D(1, 1, 7), Point3D(4, 3, 9))
+    >>> s.points
+    (Point3D(1, 1, 7), Point3D(4, 3, 9))
+    >>> s.length
+    sqrt(17)
+    >>> s.midpoint
+    Point3D(5/2, 2, 8)
 
     """
+    def __new__(cls, p1, *pts, **kwargs):
+        if isinstance(p1, LinearEntity):
+            ambient_dim = len(p1.p1)
+        else:
+            ambient_dim = len(p1)
 
-    def __new__(cls, p1, p2, **kwargs):
-        # Reorder the two points under the following ordering:
-        #   if p1.x != p2.x then p1.x < p2.x
-        #   if p1.x == p2.x then p1.y < p2.y
-        p1 = Point(p1)
-        p2 = Point(p2)
-        if p1 == p2:
-            return Point(p1)
-        if (p1.x > p2.x) == True:
-            p1, p2 = p2, p1
-        elif (p1.x == p2.x) == True and (p1.y > p2.y) == True:
-            p1, p2 = p2, p1
-        return LinearEntity.__new__(cls, p1, p2, **kwargs)
+        if ambient_dim == 2:
+            return Segment2D(p1, *pts, **kwargs)
+        elif ambient_dim == 3:
+            return Segment3D(p1, *pts, **kwargs)
+        return LinearEntity.__new__(cls, p1, *pts, **kwargs)
 
     def plot_interval(self, parameter='t'):
         """The plot interval for the default geometric plot of the Segment gives
@@ -1595,13 +1517,13 @@ class Segment(LinearEntity):
         >>> p1, p2, p3 = Point(0, 0), Point(6, 6), Point(5, 1)
         >>> s1 = Segment(p1, p2)
         >>> s1.perpendicular_bisector()
-        Line(Point2D(3, 3), Point2D(-3, 9))
+        Line2D(Point2D(3, 3), Point2D(-3, 9))
 
         >>> s1.perpendicular_bisector(p3)
-        Segment(Point2D(3, 3), Point2D(5, 1))
+        Segment2D(Point2D(3, 3), Point2D(5, 1))
 
         """
-        l = LinearEntity.perpendicular_line(self, self.midpoint)
+        l = self.perpendicular_line(self.midpoint)
         if p is None or Point(p) not in l:
             return l
         else:
@@ -1624,6 +1546,11 @@ class Segment(LinearEntity):
         >>> s1 = Segment(p1, p2)
         >>> s1.length
         5
+        >>> from sympy import Point3D, Segment3D
+        >>> p1, p2 = Point3D(0, 0, 0), Point3D(4, 3, 3)
+        >>> s1 = Segment3D(p1, p2)
+        >>> s1.length
+        sqrt(34)
 
         """
         return Point.distance(self.p1, self.p2)
@@ -1645,6 +1572,11 @@ class Segment(LinearEntity):
         >>> s1 = Segment(p1, p2)
         >>> s1.midpoint
         Point2D(2, 3/2)
+        >>> from sympy import Point3D, Segment3D
+        >>> p1, p2 = Point3D(0, 0, 0), Point3D(4, 3, 3)
+        >>> s1 = Segment3D(p1, p2)
+        >>> s1.midpoint
+        Point3D(2, 3/2, 3/2)
 
         """
         return Point.midpoint(self.p1, self.p2)
@@ -1656,7 +1588,7 @@ class Segment(LinearEntity):
         Raises
         ======
 
-        NotImplementedError is raised if o is not a Point
+        NotImplementedError is raised if `other` is not a Point
 
         Examples
         ========
@@ -1668,6 +1600,13 @@ class Segment(LinearEntity):
         sqrt(170)
         >>> s.distance((0, 12))
         sqrt(73)
+        >>> from sympy import Point3D, Segment3D
+        >>> p1, p2 = Point3D(0, 0, 3), Point3D(1, 1, 4)
+        >>> s = Segment3D(p1, p2)
+        >>> s.distance(Point3D(10, 15, 12))
+        sqrt(341)
+        >>> s.distance((10, 15, 12))
+        sqrt(341)
         """
         if is_sequence(other):
             other = Point(other)
@@ -1698,10 +1637,24 @@ class Segment(LinearEntity):
         >>> s2 = Segment(p2, p1)
         >>> s.contains(s2)
         True
+        >>> from sympy import Point3D, Segment3D
+        >>> p1, p2 = Point3D(0, 1, 1), Point3D(3, 4, 5)
+        >>> s = Segment3D(p1, p2)
+        >>> s2 = Segment3D(p2, p1)
+        >>> s.contains(s2)
+        True
+        >>> s.contains((p1 + p2) / 2)
+        True
         """
 
         if is_sequence(other):
             other = Point(other)
+
+        if hasattr(other, 'ambient_dimension') and other.ambient_dimension != self.ambient_dimension:
+            warnings.warn("Testing containment for objects of different ambient dimensions", RuntimeWarning)
+
+            raise Exception()
+            return False
 
         if isinstance(other, Point):
             if Point.is_collinear(other, self.p1, self.p2):
@@ -1719,6 +1672,466 @@ class Segment(LinearEntity):
             return other.p1 in self and other.p2 in self
 
         return False
+
+class LinearEntity2D(LinearEntity):
+    """A base class for all linear entities (line, ray and segment)
+    in a 2-dimensional Euclidean space.
+
+    Attributes
+    ==========
+
+    p1
+    p2
+    coefficients
+    slope
+    points
+
+    Notes
+    =====
+
+    This is an abstract class and is not meant to be instantiated.
+
+    See Also
+    ========
+
+    sympy.geometry.entity.GeometryEntity
+
+    """
+    @property
+    def bounds(self):
+        """Return a tuple (xmin, ymin, xmax, ymax) representing the bounding
+        rectangle for the geometric figure.
+
+        """
+
+        verts = self.points
+        xs = [p.x for p in verts]
+        ys = [p.y for p in verts]
+        return (min(xs), min(ys), max(xs), max(ys))
+
+    @property
+    def slope(self):
+        """The slope of this linear entity, or infinity if vertical.
+
+        Returns
+        =======
+
+        slope : number or sympy expression
+
+        See Also
+        ========
+
+        coefficients
+
+        Examples
+        ========
+
+        >>> from sympy import Point, Line
+        >>> p1, p2 = Point(0, 0), Point(3, 5)
+        >>> l1 = Line(p1, p2)
+        >>> l1.slope
+        5/3
+
+        >>> p3 = Point(0, 4)
+        >>> l2 = Line(p1, p3)
+        >>> l2.slope
+        oo
+
+        """
+        d1, d2 = (self.p1 - self.p2).args
+        if d1 == 0:
+            return S.Infinity
+        return simplify(d2/d1)
+
+    def perpendicular_line(self, p):
+        """Create a new Line perpendicular to this linear entity which passes
+        through the point `p`.
+
+        Parameters
+        ==========
+
+        p : Point
+
+        Returns
+        =======
+
+        line : Line
+
+        See Also
+        ========
+
+        is_perpendicular, perpendicular_segment
+
+        Examples
+        ========
+
+        >>> from sympy import Point, Line
+        >>> p1, p2, p3 = Point(0, 0), Point(2, 3), Point(-2, 2)
+        >>> l1 = Line(p1, p2)
+        >>> l2 = l1.perpendicular_line(p3)
+        >>> p3 in l2
+        True
+        >>> l1.is_perpendicular(l2)
+        True
+
+        """
+        p = Point(p)
+        # any two lines in R^2 intersect, so blindly making
+        # a line through p in an orthogonal direction will work
+        return Line(p, p + self.direction.orthogonal_direction)
+
+class Line2D(LinearEntity2D, Line):
+    """An infinite line in space 2D.
+
+    A line is declared with two distinct points or a point and slope
+    as defined using keyword `slope`.
+
+    Parameters
+    ==========
+
+    p1 : Point
+    pt : Point
+    slope : sympy expression
+
+    See Also
+    ========
+
+    sympy.geometry.point.Point
+
+    Examples
+    ========
+
+    >>> import sympy
+    >>> from sympy import Point
+    >>> from sympy.abc import L
+    >>> from sympy.geometry import Line, Segment
+    >>> L = Line(Point(2,3), Point(3,5))
+    >>> L
+    Line2D(Point2D(2, 3), Point2D(3, 5))
+    >>> L.points
+    (Point2D(2, 3), Point2D(3, 5))
+    >>> L.equation()
+    -2*x + y + 1
+    >>> L.coefficients
+    (-2, 1, 1)
+
+    Instantiate with keyword ``slope``:
+
+    >>> Line(Point(0, 0), slope=0)
+    Line2D(Point2D(0, 0), Point2D(1, 0))
+
+    Instantiate with another linear object
+
+    >>> s = Segment((0, 0), (0, 1))
+    >>> Line(s).equation()
+    x
+    """
+    def __new__(cls, p1, pt=None, slope=None, **kwargs):
+        if isinstance(p1, LinearEntity):
+            p1, pt = p1.args
+        else:
+            p1 = Point(p1)
+        if pt is not None and slope is None:
+            try:
+                p2 = Point(pt)
+            except NotImplementedError:
+                raise ValueError('The 2nd argument was not a valid Point. '
+                'If it was a slope, enter it with keyword "slope".')
+        elif slope is not None and pt is None:
+            slope = sympify(slope)
+            if slope.is_finite is False:
+                # when infinite slope, don't change x
+                dx = 0
+                dy = 1
+            else:
+                # go over 1 up slope
+                dx = 1
+                dy = slope
+            # XXX avoiding simplification by adding to coords directly
+            p2 = Point(p1.x + dx, p1.y + dy)
+        else:
+            raise ValueError('A 2nd Point or keyword "slope" must be used.')
+        return LinearEntity2D.__new__(cls, p1, p2, **kwargs)
+
+    def equation(self, x='x', y='y'):
+        """The equation of the line: ax + by + c.
+
+        Parameters
+        ==========
+
+        x : str, optional
+            The name to use for the x-axis, default value is 'x'.
+        y : str, optional
+            The name to use for the y-axis, default value is 'y'.
+
+        Returns
+        =======
+
+        equation : sympy expression
+
+        See Also
+        ========
+
+        LinearEntity.coefficients
+
+        Examples
+        ========
+
+        >>> from sympy import Point, Line
+        >>> p1, p2 = Point(1, 0), Point(5, 3)
+        >>> l1 = Line(p1, p2)
+        >>> l1.equation()
+        -3*x + 4*y + 3
+
+        """
+        x, y = _symbol(x), _symbol(y)
+        p1, p2 = self.points
+        if p1.x == p2.x:
+            return x - p1.x
+        elif p1.y == p2.y:
+            return y - p1.y
+
+        a, b, c = self.coefficients
+        return a*x + b*y + c
+
+    def _svg(self, scale_factor=1., fill_color="#66cc99"):
+        """Returns SVG path element for the LinearEntity.
+
+        Parameters
+        ==========
+
+        scale_factor : float
+            Multiplication factor for the SVG stroke-width.  Default is 1.
+        fill_color : str, optional
+            Hex string for fill color. Default is "#66cc99".
+        """
+
+        from sympy.core.evalf import N
+
+        verts = (N(self.p1), N(self.p2))
+        coords = ["{0},{1}".format(p.x, p.y) for p in verts]
+        path = "M {0} L {1}".format(coords[0], " L ".join(coords[1:]))
+
+        return (
+            '<path fill-rule="evenodd" fill="{2}" stroke="#555555" '
+            'stroke-width="{0}" opacity="0.6" d="{1}" '
+            'marker-start="url(#markerReverseArrow)" marker-end="url(#markerArrow)"/>'
+            ).format(2. * scale_factor, path, fill_color)
+
+class Ray2D(LinearEntity2D, Ray):
+    """
+    A Ray is a semi-line in the space with a source point and a direction.
+
+    Parameters
+    ==========
+
+    p1 : Point
+        The source of the Ray
+    p2 : Point or radian value
+        This point determines the direction in which the Ray propagates.
+        If given as an angle it is interpreted in radians with the positive
+        direction being ccw.
+
+    Attributes
+    ==========
+
+    source
+    xdirection
+    ydirection
+
+    See Also
+    ========
+
+    sympy.geometry.point.Point, Line
+
+    Examples
+    ========
+
+    >>> import sympy
+    >>> from sympy import Point, pi
+    >>> from sympy.abc import r
+    >>> from sympy.geometry import Ray
+    >>> r = Ray(Point(2, 3), Point(3, 5))
+    >>> r = Ray(Point(2, 3), Point(3, 5))
+    >>> r
+    Ray2D(Point2D(2, 3), Point2D(3, 5))
+    >>> r.points
+    (Point2D(2, 3), Point2D(3, 5))
+    >>> r.source
+    Point2D(2, 3)
+    >>> r.xdirection
+    oo
+    >>> r.ydirection
+    oo
+    >>> r.slope
+    2
+    >>> Ray(Point(0, 0), angle=pi/4).slope
+    1
+
+    """
+    def __new__(cls, p1, pt=None, angle=None, **kwargs):
+        p1 = Point(p1)
+        if pt is not None and angle is None:
+            try:
+                p2 = Point(pt)
+            except NotImplementedError:
+                from sympy.utilities.misc import filldedent
+                raise ValueError(filldedent('''
+                    The 2nd argument was not a valid Point; if
+                    it was meant to be an angle it should be
+                    given with keyword "angle".'''))
+            if p1 == p2:
+                raise ValueError('A Ray requires two distinct points.')
+        elif angle is not None and pt is None:
+            # we need to know if the angle is an odd multiple of pi/2
+            c = pi_coeff(sympify(angle))
+            p2 = None
+            if c is not None:
+                if c.is_Rational:
+                    if c.q == 2:
+                        if c.p == 1:
+                            p2 = p1 + Point(0, 1)
+                        elif c.p == 3:
+                            p2 = p1 + Point(0, -1)
+                    elif c.q == 1:
+                        if c.p == 0:
+                            p2 = p1 + Point(1, 0)
+                        elif c.p == 1:
+                            p2 = p1 + Point(-1, 0)
+                if p2 is None:
+                    c *= S.Pi
+            else:
+                c = angle % (2*S.Pi)
+            if not p2:
+                m = 2*c/S.Pi
+                left = And(1 < m, m < 3)  # is it in quadrant 2 or 3?
+                x = Piecewise((-1, left), (Piecewise((0, Eq(m % 1, 0)), (1, True)), True))
+                y = Piecewise((-tan(c), left), (Piecewise((1, Eq(m, 1)), (-1, Eq(m, 3)), (tan(c), True)), True))
+                p2 = p1 + Point(x, y)
+        else:
+            raise ValueError('A 2nd point or keyword "angle" must be used.')
+
+        return LinearEntity2D.__new__(cls, p1, p2, **kwargs)
+
+    @property
+    def xdirection(self):
+        """The x direction of the ray.
+
+        Positive infinity if the ray points in the positive x direction,
+        negative infinity if the ray points in the negative x direction,
+        or 0 if the ray is vertical.
+
+        See Also
+        ========
+
+        ydirection
+
+        Examples
+        ========
+
+        >>> from sympy import Point, Ray
+        >>> p1, p2, p3 = Point(0, 0), Point(1, 1), Point(0, -1)
+        >>> r1, r2 = Ray(p1, p2), Ray(p1, p3)
+        >>> r1.xdirection
+        oo
+        >>> r2.xdirection
+        0
+
+        """
+        if self.p1.x < self.p2.x:
+            return S.Infinity
+        elif self.p1.x == self.p2.x:
+            return S.Zero
+        else:
+            return S.NegativeInfinity
+
+    @property
+    def ydirection(self):
+        """The y direction of the ray.
+
+        Positive infinity if the ray points in the positive y direction,
+        negative infinity if the ray points in the negative y direction,
+        or 0 if the ray is horizontal.
+
+        See Also
+        ========
+
+        xdirection
+
+        Examples
+        ========
+
+        >>> from sympy import Point, Ray
+        >>> p1, p2, p3 = Point(0, 0), Point(-1, -1), Point(-1, 0)
+        >>> r1, r2 = Ray(p1, p2), Ray(p1, p3)
+        >>> r1.ydirection
+        -oo
+        >>> r2.ydirection
+        0
+
+        """
+        if self.p1.y < self.p2.y:
+            return S.Infinity
+        elif self.p1.y == self.p2.y:
+            return S.Zero
+        else:
+            return S.NegativeInfinity
+
+class Segment2D(LinearEntity2D, Segment):
+    """An undirected line segment in 2D space.
+
+    Parameters
+    ==========
+
+    p1 : Point
+    p2 : Point
+
+    Attributes
+    ==========
+
+    length : number or sympy expression
+    midpoint : Point
+
+    See Also
+    ========
+
+    sympy.geometry.point.Point, Line
+
+    Examples
+    ========
+
+    >>> import sympy
+    >>> from sympy import Point
+    >>> from sympy.abc import s
+    >>> from sympy.geometry import Segment
+    >>> Segment((1, 0), (1, 1)) # tuples are interpreted as pts
+    Segment2D(Point2D(1, 0), Point2D(1, 1))
+    >>> s = Segment(Point(4, 3), Point(1, 1))
+    >>> s
+    Segment2D(Point2D(1, 1), Point2D(4, 3))
+    >>> s.points
+    (Point2D(1, 1), Point2D(4, 3))
+    >>> s.slope
+    2/3
+    >>> s.length
+    sqrt(13)
+    >>> s.midpoint
+    Point2D(5/2, 2)
+
+    """
+    def __new__(cls, p1, p2, **kwargs):
+        # Reorder the two points under the following ordering:
+        #   if p1.x != p2.x then p1.x < p2.x
+        #   if p1.x == p2.x then p1.y < p2.y
+        p1 = Point(p1)
+        p2 = Point(p2)
+        if p1 == p2:
+            return Point(p1)
+        if (p1.x > p2.x) == True:
+            p1, p2 = p2, p1
+        elif (p1.x == p2.x) == True and (p1.y > p2.y) == True:
+            p1, p2 = p2, p1
+        return LinearEntity2D.__new__(cls, p1, p2, **kwargs)
 
     def _svg(self, scale_factor=1., fill_color="#66cc99"):
         """Returns SVG path element for the LinearEntity.
@@ -1741,3 +2154,382 @@ class Segment(LinearEntity):
             '<path fill-rule="evenodd" fill="{2}" stroke="#555555" '
             'stroke-width="{0}" opacity="0.6" d="{1}" />'
             ).format(2. * scale_factor, path, fill_color)
+
+class LinearEntity3D(LinearEntity):
+    """An base class for all linear entities (line, ray and segment)
+    in a 3-dimensional Euclidean space.
+
+    Attributes
+    ==========
+
+    p1
+    p2
+    direction_ratio
+    direction_cosine
+    points
+
+    Notes
+    =====
+
+    This is a base class and is not meant to be instantiated.
+    """
+
+    def __new__(cls, p1, p2, **kwargs):
+        p1 = Point3D(p1)
+        p2 = Point3D(p2)
+        if p1 == p2:
+            # if it makes sense to return a Point, handle in subclass
+            raise ValueError(
+                "%s.__new__ requires two unique Points." % cls.__name__)
+
+        return GeometryEntity.__new__(cls, p1, p2, **kwargs)
+
+    ambient_dimension = 3
+
+    @property
+    def direction_ratio(self):
+        """The direction ratio of a given line in 3D.
+
+        See Also
+        ========
+
+        sympy.geometry.line.Line.equation
+
+        Examples
+        ========
+
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2 = Point3D(0, 0, 0), Point3D(5, 3, 1)
+        >>> l = Line3D(p1, p2)
+        >>> l.direction_ratio
+        [5, 3, 1]
+        """
+        p1, p2 = self.points
+        return p1.direction_ratio(p2)
+
+    @property
+    def direction_cosine(self):
+        """The normalized direction ratio of a given line in 3D.
+
+        See Also
+        ========
+
+        sympy.geometry.line.Line.equation
+
+        Examples
+        ========
+
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2 = Point3D(0, 0, 0), Point3D(5, 3, 1)
+        >>> l = Line3D(p1, p2)
+        >>> l.direction_cosine
+        [sqrt(35)/7, 3*sqrt(35)/35, sqrt(35)/35]
+        >>> sum(i**2 for i in _)
+        1
+        """
+        p1, p2 = self.points
+        return p1.direction_cosine(p2)
+
+class Line3D(LinearEntity3D, Line):
+    """An infinite 3D line in space.
+
+    A line is declared with two distinct points or a point and direction_ratio
+    as defined using keyword `direction_ratio`.
+
+    Parameters
+    ==========
+
+    p1 : Point3D
+    pt : Point3D
+    direction_ratio : list
+
+    See Also
+    ========
+
+    sympy.geometry.point.Point3D
+    sympy.geometry.line.Line
+    sympy.geometry.line.Line2D
+
+    Examples
+    ========
+
+    >>> import sympy
+    >>> from sympy import Point3D
+    >>> from sympy.abc import L
+    >>> from sympy.geometry import Line3D, Segment3D
+    >>> L = Line3D(Point3D(2, 3, 4), Point3D(3, 5, 1))
+    >>> L
+    Line3D(Point3D(2, 3, 4), Point3D(3, 5, 1))
+    >>> L.points
+    (Point3D(2, 3, 4), Point3D(3, 5, 1))
+    """
+
+    def __new__(cls, p1, pt=None, direction_ratio=[], **kwargs):
+        if isinstance(p1, LinearEntity3D):
+            p1, pt = p1.args
+        else:
+            p1 = Point3D(p1)
+        if pt is not None and len(direction_ratio) == 0:
+            pt = Point3D(pt)
+        elif len(direction_ratio) == 3 and pt is None:
+            pt = Point3D(p1.x + direction_ratio[0], p1.y + direction_ratio[1],
+                         p1.z + direction_ratio[2])
+        else:
+            raise ValueError('A 2nd Point or keyword "direction_ratio" must '
+            'be used.')
+
+        return LinearEntity3D.__new__(cls, p1, pt, **kwargs)
+
+    def equation(self, x='x', y='y', z='z', k='k'):
+        """The equation of the line in 3D
+
+        Parameters
+        ==========
+
+        x : str, optional
+            The name to use for the x-axis, default value is 'x'.
+        y : str, optional
+            The name to use for the y-axis, default value is 'y'.
+        z : str, optional
+            The name to use for the x-axis, default value is 'z'.
+
+        Returns
+        =======
+
+        equation : tuple
+
+        Examples
+        ========
+
+        >>> from sympy import Point3D, Line3D
+        >>> p1, p2 = Point3D(1, 0, 0), Point3D(5, 3, 0)
+        >>> l1 = Line3D(p1, p2)
+        >>> l1.equation()
+        (x/4 - 1/4, y/3, zoo*z, k)
+
+        """
+        x, y, z, k = _symbol(x), _symbol(y), _symbol(z), _symbol(k)
+        p1, p2 = self.points
+        a = p1.direction_ratio(p2)
+        return (((x - p1.x)/a[0]), ((y - p1.y)/a[1]),
+                ((z - p1.z)/a[2]), k)
+
+class Ray3D(LinearEntity3D, Ray):
+    """
+    A Ray is a semi-line in the space with a source point and a direction.
+
+    Parameters
+    ==========
+
+    p1 : Point3D
+        The source of the Ray
+    p2 : Point or a direction vector
+    direction_ratio: Determines the direction in which the Ray propagates.
+
+
+    Attributes
+    ==========
+
+    source
+    xdirection
+    ydirection
+    zdirection
+
+    See Also
+    ========
+
+    sympy.geometry.point.Point3D, Line3D
+
+
+    Examples
+    ========
+
+    >>> import sympy
+    >>> from sympy import Point3D, pi
+    >>> from sympy.abc import r
+    >>> from sympy.geometry import Ray3D
+    >>> r = Ray3D(Point3D(2, 3, 4), Point3D(3, 5, 0))
+    >>> r
+    Ray3D(Point3D(2, 3, 4), Point3D(3, 5, 0))
+    >>> r.points
+    (Point3D(2, 3, 4), Point3D(3, 5, 0))
+    >>> r.source
+    Point3D(2, 3, 4)
+    >>> r.xdirection
+    oo
+    >>> r.ydirection
+    oo
+    >>> r.direction_ratio
+    [1, 2, -4]
+
+    """
+
+    def __new__(cls, p1, pt=None, direction_ratio=[], **kwargs):
+        if isinstance(p1, LinearEntity3D):
+            p1, pt = p1.args
+        else:
+            p1 = Point3D(p1)
+        if pt is not None and len(direction_ratio) == 0:
+            pt = Point3D(pt)
+        elif len(direction_ratio) == 3 and pt is None:
+            pt = Point3D(p1.x + direction_ratio[0], p1.y + direction_ratio[1],
+                         p1.z + direction_ratio[2])
+        else:
+            raise ValueError('A 2nd Point or keyword "direction_ratio" must'
+            'be used.')
+
+        return LinearEntity3D.__new__(cls, p1, pt, **kwargs)
+
+    @property
+    def xdirection(self):
+        """The x direction of the ray.
+
+        Positive infinity if the ray points in the positive x direction,
+        negative infinity if the ray points in the negative x direction,
+        or 0 if the ray is vertical.
+
+        See Also
+        ========
+
+        ydirection
+
+        Examples
+        ========
+
+        >>> from sympy import Point3D, Ray3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(1, 1, 1), Point3D(0, -1, 0)
+        >>> r1, r2 = Ray3D(p1, p2), Ray3D(p1, p3)
+        >>> r1.xdirection
+        oo
+        >>> r2.xdirection
+        0
+
+        """
+        if self.p1.x < self.p2.x:
+            return S.Infinity
+        elif self.p1.x == self.p2.x:
+            return S.Zero
+        else:
+            return S.NegativeInfinity
+
+    @property
+    def ydirection(self):
+        """The y direction of the ray.
+
+        Positive infinity if the ray points in the positive y direction,
+        negative infinity if the ray points in the negative y direction,
+        or 0 if the ray is horizontal.
+
+        See Also
+        ========
+
+        xdirection
+
+        Examples
+        ========
+
+        >>> from sympy import Point3D, Ray3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(-1, -1, -1), Point3D(-1, 0, 0)
+        >>> r1, r2 = Ray3D(p1, p2), Ray3D(p1, p3)
+        >>> r1.ydirection
+        -oo
+        >>> r2.ydirection
+        0
+
+        """
+        if self.p1.y < self.p2.y:
+            return S.Infinity
+        elif self.p1.y == self.p2.y:
+            return S.Zero
+        else:
+            return S.NegativeInfinity
+
+    @property
+    def zdirection(self):
+        """The z direction of the ray.
+
+        Positive infinity if the ray points in the positive z direction,
+        negative infinity if the ray points in the negative z direction,
+        or 0 if the ray is horizontal.
+
+        See Also
+        ========
+
+        xdirection
+
+        Examples
+        ========
+
+        >>> from sympy import Point3D, Ray3D
+        >>> p1, p2, p3 = Point3D(0, 0, 0), Point3D(-1, -1, -1), Point3D(-1, 0, 0)
+        >>> r1, r2 = Ray3D(p1, p2), Ray3D(p1, p3)
+        >>> r1.ydirection
+        -oo
+        >>> r2.ydirection
+        0
+        >>> r2.zdirection
+        0
+
+        """
+        if self.p1.z < self.p2.z:
+            return S.Infinity
+        elif self.p1.z == self.p2.z:
+            return S.Zero
+        else:
+            return S.NegativeInfinity
+
+class Segment3D(LinearEntity3D, Segment):
+    """A undirected line segment in a 3D space.
+
+    Parameters
+    ==========
+
+    p1 : Point3D
+    p2 : Point3D
+
+    Attributes
+    ==========
+
+    length : number or sympy expression
+    midpoint : Point3D
+
+    See Also
+    ========
+
+    sympy.geometry.point.Point3D, Line3D
+
+    Examples
+    ========
+
+    >>> import sympy
+    >>> from sympy import Point3D
+    >>> from sympy.abc import s
+    >>> from sympy.geometry import Segment3D
+    >>> Segment3D((1, 0, 0), (1, 1, 1)) # tuples are interpreted as pts
+    Segment3D(Point3D(1, 0, 0), Point3D(1, 1, 1))
+    >>> s = Segment3D(Point3D(4, 3, 9), Point3D(1, 1, 7))
+    >>> s
+    Segment3D(Point3D(1, 1, 7), Point3D(4, 3, 9))
+    >>> s.points
+    (Point3D(1, 1, 7), Point3D(4, 3, 9))
+    >>> s.length
+    sqrt(17)
+    >>> s.midpoint
+    Point3D(5/2, 2, 8)
+
+    """
+
+    def __new__(cls, p1, p2, **kwargs):
+        # Reorder the two points under the following ordering:
+        #   if p1.x != p2.x then p1.x < p2.x
+        #   if p1.x == p2.x then p1.y < p2.y
+        #   The z-coordinate will not come into picture while ordering
+        p1 = Point3D(p1)
+        p2 = Point3D(p2)
+        if p1 == p2:
+            return Point3D(p1)
+        if (p1.x > p2.x) == True:
+            p1, p2 = p2, p1
+        elif (p1.x == p2.x) == True and (p1.y > p2.y) == True:
+            p1, p2 = p2, p1
+        return LinearEntity3D.__new__(cls, p1, p2, **kwargs)
