+++ sympy/geometry/tests/test_line.py
@@ -10,6 +10,8 @@ from sympy.geometry.polygon import _asa as asa
 from sympy.utilities.iterables import cartes
 from sympy.utilities.pytest import raises
 
+import warnings
+
 
 def feq(a, b):
     """Test if two floating point values are 'equal'."""
@@ -372,7 +374,7 @@ def test_line3d():
     l1_1 = Line3D(p1, p1_1)
     assert Line3D.is_perpendicular(l1, l2) is False
     p = l1.arbitrary_point()
-    raises(NotImplementedError , lambda: l1.perpendicular_segment(p))
+    assert l1.perpendicular_segment(p) == p
 
     # Parallelity
     assert l1.parallel_line(p1_1) == Line3D(Point3D(x1, x1, x1),
@@ -391,7 +393,7 @@ def test_line3d():
     assert line3.intersection(line4) == [Point3D(2, 2, 1)]
     assert line3.is_parallel(line4) is False
     assert Line3D((0, 1, 2), (0, 2, 3)).intersection(
-        Line3D((0, 1, 2), (0, 1, 1))) == []
+        Line3D((0, 1, 2), (0, 1, 1))) == [Point3D(0,1,2)]
     ray0 = Ray3D((0, 0), (3, 0))
     ray1 = Ray3D((1, 0), (3, 0))
     assert ray0.intersection(ray1) == [ray1]
@@ -438,10 +440,8 @@ def test_line3d():
     r3 = Ray3D(p1, p2)
     r4 = Ray3D(p2, p1)
     r5 = Ray3D(Point3D(0, 1, 1), Point3D(1, 2, 0))
-    assert l1.projection(r1) == [
-        Ray3D(Point3D(0, 0, 0), Point3D(4/3, 4/3, 4/3))]
-    assert l1.projection(r2) == [
-        Ray3D(Point3D(0, 0, 0), Point3D(1/3, 1/3, 1/3))]
+    assert l1.projection(r1) == Ray3D(Point3D(0, 0, 0), Point3D(4/3, 4/3, 4/3))
+    assert l1.projection(r2) == Ray3D(Point3D(0, 0, 0), Point3D(1/3, 1/3, 1/3))
     assert r3 != r1
     t = Symbol('t', real=True)
     assert Ray3D((1, 1, 1), direction_ratio=[1, 2, 3]).arbitrary_point() == \
@@ -471,7 +471,7 @@ def test_line3d():
     pt2 = Point3D(Rational(3)/2, Rational(3)/2, Rational(3)/2)
     assert s1.distance(pt1) == 0
     assert s2.distance(pt1) == sqrt(3)/2
-    assert s2.distance(pt2) == 2
+    assert s2.distance(pt2) == 2*sqrt(6)/3
     assert s1.distance((0,0,0)) == 0
     assert s2.distance((0,0,0)) == sqrt(3)/2
     # Line to point
@@ -492,8 +492,11 @@ def test_line3d():
     assert r.distance(Point3D(1, 1, 1)) == 0
     assert r.distance((-1, -1, -1)) == sqrt(3)
     assert r.distance((1, 1, 1)) == 0
+    assert Ray3D((0,0,0),(1,1,2)).distance((-1,-1,2)) == 4*sqrt(3)/3
+    assert Ray3D((1, 1, 1), (2, 2, 2)).distance(Point3D(1.5, -3, -1)) == \
+        Rational(9)/2
     assert Ray3D((1, 1, 1), (2, 2, 2)).distance(Point3D(1.5, 3, 1)) == \
-        sqrt(17)/2
+        sqrt(78)/6
 
 
     # Special cases of projection and intersection
@@ -525,9 +528,9 @@ def test_line3d():
     assert intersection(s1, r1) == [s1]
 
     # check that temporary symbol is Dummy
-    assert Line3D((0, 0), (t, t)).perpendicular_line((0, 1)).equals( \
+    assert Line3D((0, 0), (t, t)).perpendicular_line((0, 1, 0)).equals( \
         Line3D(Point3D(0, 1, 0), Point3D(1/2, 1/2, 0)))
-    assert Line3D((0, 0), (t, t)).perpendicular_segment((0, 1)).equals( \
+    assert Line3D((0, 0), (t, t)).perpendicular_segment((0, 1, 0)).equals( \
         Segment3D(Point3D(0, 1, 0), Point3D(1/2, 1/2, 0)))
     assert Line3D((0, 0), (t, t)).intersection(Line3D((0, 1), (t, t))) == \
         [Point3D(t, t, 0)]
@@ -540,7 +543,7 @@ def test_line3d():
 
     # Test projection
     assert parallel_1.projection(Point3D(5, 5, 0)) == Point3D(5, 0, 0)
-    assert parallel_1.projection(parallel_2) == [parallel_1]
+    assert parallel_1.projection(parallel_2).equals( parallel_1 )
     raises(GeometryError, lambda: parallel_1.projection(Plane(p1, p2, p6)))
 
     # Test __new__
@@ -549,7 +552,9 @@ def test_line3d():
 
     # Test contains
     pt2d = Point(1.0, 1.0)
-    assert perp_1.contains(pt2d) is False
+    with warnings.catch_warnings(record=True) as w:
+        assert perp_1.contains(pt2d) is False
+        assert len(w) == 1
 
     # Test equals
     assert perp_1.equals(pt2d) is False
@@ -574,9 +579,11 @@ def test_line3d():
     posz = Ray3D(p1, Point3D(0, 0, 1))
     assert posy.contains(p1) is True
     assert posz.contains(p1) is True
-    assert posz.contains(pt2d) is False
+    with warnings.catch_warnings(record=True) as w:
+        assert posz.contains(pt2d) is False
+        assert len(w) == 1
     ray1 = Ray3D(Point3D(1, 1, 1), Point3D(1, 0, 0))
-    raises(TypeError, lambda: ray1.contains([]))
+    raises(ValueError, lambda: ray1.contains([]))
 
     # Test equals
     assert negz.equals(pt2d) is False
@@ -584,11 +591,11 @@ def test_line3d():
 
     assert ray1.is_similar(Line3D(Point3D(1, 1, 1), Point3D(1, 0, 0))) is True
     assert ray1.is_similar(perp_1) is False
-    raises(NotImplementedError, lambda: ray1.is_similar(ray1))
+    assert ray1.is_similar(ray1) is True
 
     # Begin Segment
     seg1 = Segment3D(p1, Point3D(1, 0, 0))
-    raises(TypeError, lambda: seg1.contains([]))
+    raises(ValueError, lambda: seg1.contains([]))
     seg2= Segment3D(Point3D(2, 2, 2), Point3D(3, 2, 2))
     assert seg1.contains(seg2) is False
